pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Viraat", email = "viraat.laldas@gmail.com"},
]
dependencies = [
    "python-chess~=0.31",
]
requires-python = ">=3.6"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
